Move the header to be part of the grid
This makes it possible to scroll with the grid content so that a
scrollbar doesn't end in the middle of the window. This is also better
on mobile since the header take a lot of space.

Other than that this contains various small cleanup to make the QML code
nicer.